# pb/rpc_verify_email.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.36.6
// 	protoc        v5.29.3
// source: rpc_verify_email.proto

package pb

import (
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	reflect "reflect"
	sync "sync"
	unsafe "unsafe"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

type CreateVerifyEmailRequest struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	EmailId       int64                  `protobuf:"varint,1,opt,name=email_id,json=emailId,proto3" json:"email_id,omitempty"`
	SecretCode    string                 `protobuf:"bytes,2,opt,name=secret_code,json=secretCode,proto3" json:"secret_code,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *CreateVerifyEmailRequest) Reset() {
	*x = CreateVerifyEmailRequest{}
	mi := &file_rpc_verify_email_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *CreateVerifyEmailRequest) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*CreateVerifyEmailRequest) ProtoMessage() {}

func (x *CreateVerifyEmailRequest) ProtoReflect() protoreflect.Message {
	mi := &file_rpc_verify_email_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use CreateVerifyEmailRequest.ProtoReflect.Descriptor instead.
func (*CreateVerifyEmailRequest) Descriptor() ([]byte, []int) {
	return file_rpc_verify_email_proto_rawDescGZIP(), []int{0}
}

func (x *CreateVerifyEmailRequest) GetEmailId() int64 {
	if x != nil {
		return x.EmailId
	}
	return 0
}

func (x *CreateVerifyEmailRequest) GetSecretCode() string {
	if x != nil {
		return x.SecretCode
	}
	return ""
}

type CreateVerifyEmailResponse struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	IsVerified    bool                   `protobuf:"varint,1,opt,name=is_verified,json=isVerified,proto3" json:"is_verified,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *CreateVerifyEmailResponse) Reset() {
	*x = CreateVerifyEmailResponse{}
	mi := &file_rpc_verify_email_proto_msgTypes[1]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *CreateVerifyEmailResponse) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*CreateVerifyEmailResponse) ProtoMessage() {}

func (x *CreateVerifyEmailResponse) ProtoReflect() protoreflect.Message {
	mi := &file_rpc_verify_email_proto_msgTypes[1]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use CreateVerifyEmailResponse.ProtoReflect.Descriptor instead.
func (*CreateVerifyEmailResponse) Descriptor() ([]byte, []int) {
	return file_rpc_verify_email_proto_rawDescGZIP(), []int{1}
}

func (x *CreateVerifyEmailResponse) GetIsVerified() bool {
	if x != nil {
		return x.IsVerified
	}
	return false
}

var File_rpc_verify_email_proto protoreflect.FileDescriptor

const file_rpc_verify_email_proto_rawDesc = "" +
	"\n" +
	"\x16rpc_verify_email.proto\"V\n" +
	"\x18CreateVerifyEmailRequest\x12\x19\n" +
	"\bemail_id\x18\x01 \x01(\x03R\aemailId\x12\x1f\n" +
	"\vsecret_code\x18\x02 \x01(\tR\n" +
	"secretCode\"<\n" +
	"\x19CreateVerifyEmailResponse\x12\x1f\n" +
	"\vis_verified\x18\x01 \x01(\bR\n" +
	"isVerifiedB#Z!github.com/pawaspy/simple_bank/pbb\x06proto3"

var (
	file_rpc_verify_email_proto_rawDescOnce sync.Once
	file_rpc_verify_email_proto_rawDescData []byte
)

func file_rpc_verify_email_proto_rawDescGZIP() []byte {
	file_rpc_verify_email_proto_rawDescOnce.Do(func() {
		file_rpc_verify_email_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_rpc_verify_email_proto_rawDesc), len(file_rpc_verify_email_proto_rawDesc)))
	})
	return file_rpc_verify_email_proto_rawDescData
}

var file_rpc_verify_email_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
var file_rpc_verify_email_proto_goTypes = []any{
	(*CreateVerifyEmailRequest)(nil),  // 0: CreateVerifyEmailRequest
	(*CreateVerifyEmailResponse)(nil), // 1: CreateVerifyEmailResponse
}
var file_rpc_verify_email_proto_depIdxs = []int32{
	0, // [0:0] is the sub-list for method output_type
	0, // [0:0] is the sub-list for method input_type
	0, // [0:0] is the sub-list for extension type_name
	0, // [0:0] is the sub-list for extension extendee
	0, // [0:0] is the sub-list for field type_name
}

func init() { file_rpc_verify_email_proto_init() }
func file_rpc_verify_email_proto_init() {
	if File_rpc_verify_email_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: unsafe.Slice(unsafe.StringData(file_rpc_verify_email_proto_rawDesc), len(file_rpc_verify_email_proto_rawDesc)),
			NumEnums:      0,
			NumMessages:   2,
			NumExtensions: 0,
			NumServices:   0,
		},
		GoTypes:           file_rpc_verify_email_proto_goTypes,
		DependencyIndexes: file_rpc_verify_email_proto_depIdxs,
		MessageInfos:      file_rpc_verify_email_proto_msgTypes,
	}.Build()
	File_rpc_verify_email_proto = out.File
	file_rpc_verify_email_proto_goTypes = nil
	file_rpc_verify_email_proto_depIdxs = nil
}